[RENAME/#9] remote 요청, 응답 Dto 클래스명에 HTTP 메서드 이름 포함하도록 변경
