--- app/src/main/java/org/android/go/sopt/data/remote/model/Response.kt
@@ -1,73 +1,0 @@
-package org.android.go.sopt.data.remote.model
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ResSignUpDto(
-    @SerialName("status")
-    val status: Int,
-    @SerialName("message")
-    val message: String,
-    @SerialName("data")
-    val data: SignUpData,
-) {
-    @Serializable
-    data class SignUpData(
-        @SerialName("name")
-        val name: String,
-        @SerialName("skill")
-        val skill: String,
-    )
-}
-
-@Serializable
-data class ResLoginDto(
-    @SerialName("status")
-    val status: Int,
-    @SerialName("message")
-    val message: String,
-    @SerialName("data")
-    val data: LoginData,
-) {
-    @Serializable
-    data class LoginData(
-        @SerialName("id")
-        val id: String,
-        @SerialName("name")
-        val name: String,
-        @SerialName("skill")
-        val skill: String,
-    )
-}
-
-@Serializable
-data class ResFollowerDto(
-    val page: Int,
-    @SerialName("per_page")
-    val perPage: Int,
-    val total: Int,
-    @SerialName("total_pages")
-    val totalPages: Int,
-    @SerialName("data")
-    val followers: List<Follower>,
-    val support: Support
-) {
-    @Serializable
-    data class Support(
-        val url: String,
-        val text: String
-    )
-}
-
-@Serializable
-data class Follower(
-    val id: Int,
-    val email: String,
-    @SerialName("first_name")
-    val firstName: String,
-    @SerialName("last_name")
-    val lastName: String,
-    @SerialName("avatar")
-    val imgUrl: String
-)
