+++ app/src/main/java/org/android/go/sopt/data/remote/service/AuthService.kt
@@ -1,9 +1,9 @@
 package org.android.go.sopt.data.remote.service
 
-import org.android.go.sopt.data.remote.model.ReqLoginDto
-import org.android.go.sopt.data.remote.model.ReqSignUpDto
-import org.android.go.sopt.data.remote.model.ResLoginDto
-import org.android.go.sopt.data.remote.model.ResSignUpDto
+import org.android.go.sopt.data.remote.entity.request.RequestPostLoginDto
+import org.android.go.sopt.data.remote.entity.request.RequestPostSignUpDto
+import org.android.go.sopt.data.remote.entity.response.ResponsePostLoginDto
+import org.android.go.sopt.data.remote.entity.response.ResponsePostSignUpDto
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.GET
@@ -13,17 +13,17 @@ import retrofit2.http.Path
 interface AuthService {
     @POST("sign-up")
     fun signUp(
-        @Body request: ReqSignUpDto
-    ): Call<ResSignUpDto>
+        @Body request: RequestPostSignUpDto
+    ): Call<ResponsePostSignUpDto>
 
     @POST("sign-in")
     fun login(
-        @Body request: ReqLoginDto
-    ): Call<ResLoginDto>
+        @Body request: RequestPostLoginDto
+    ): Call<ResponsePostLoginDto>
 
     @GET("info/{userId}")
     fun getUserInfo(
         @Path("userId") id: String
-    ): Call<ResLoginDto>
+    ): Call<ResponsePostLoginDto>
 }
