+++ app/src/main/java/org/android/go/sopt/data/remote/service/FollowerService.kt
@@ -1,6 +1,6 @@
 package org.android.go.sopt.data.remote.service
 
-import org.android.go.sopt.data.remote.model.ResFollowerDto
+import org.android.go.sopt.data.remote.entity.response.ResponseGetFollowerListDto
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
@@ -10,5 +10,5 @@ interface FollowerService {
     fun getFollowerList(
         @Query("page") page: Int,
         @Query("per_page") perPage: Int
-    ) : Call<ResFollowerDto>
+    ) : Call<ResponseGetFollowerListDto>
 }
